requirement.txt
bcrypt==3.1.4
Django==1.11.7
Pillow==4.3.0
python-magic==0.4.13
Werkzeug==0.12.2